change: Move components from 'src' to 'app' directory
