+++ src/app/layout.tsx
@@ -1,4 +1,4 @@
-import '@/../styles/globals.css';
+import '@/styles/globals.css';
 import { Metadata } from 'next';
 import { Header } from '@/components/Header';
 import { Footer } from '@/components/Footer';
