+++ src/app/posts/types/index.ts
