+++ src/app/skills/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import { PageTitle } from '@/components/PageTitle';
-import { TechnicalSkills } from '@/components/TechnicalSkills';
+import { TechnicalSkills } from '@/app/skills/components/TechnicalSkills';
 
 const title = '技術スタック - ryo-manba';
 const description = '経験したことがある技術スタックを掲載しています。';
