--- src/components/Layout/BasicLayout.tsx
@@ -1,17 +1,0 @@
-import { ReactNode } from 'react';
-import { Header } from '@/components/Header';
-import { Footer } from '@/components/Footer';
-
-type Props = {
-  children: ReactNode;
-};
-
-export const BasicLayout = ({ children }: Props) => {
-  return (
-    <div className="flex flex-col min-h-screen">
-      <Header />
-      <main className="flex-grow">{children}</main>
-      <Footer />
-    </div>
-  );
-};
