--- src/types/NextPageWithLayout.ts
@@ -1,6 +1,0 @@
-import type { ReactElement, ReactNode } from 'react';
-import type { NextPage } from 'next';
-
-export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
-  getLayout?: (page: ReactElement) => ReactNode;
-};
